Cleaning up the code
